Add metadata to AbstractMassObject class objects

This function adds metadata to the metaData slot of an AbstractMassObject-class class object.

addMetadata(x, metadata, pos)

Arguments

x

List of AbstractMassObject-class (MassSpectra or MassPeaks) class objects.

metadata

Vector containing the metadata to be included for each element of x (same length as x).

pos

Position of the new metadata within the metaData slot list of each element of x.

Value

List of AbstractMassObject-class class objects including the new metadata in their metaData slot.

Examples

# Load example data

data(spectra) # list of MassSpectra class objects
data(type)    # metadata

# Add metadata

info <- paste("Spectrum No.",1:length(spectra)) # Artificial metadata vector

spectra2 <- addMetadata(spectra,info,1)

# Check info in metaData slot

spectra2[[1]]@metaData
#> [[1]]
#> [1] "Spectrum No. 1"
#>