Summary of mass spectra

This function generates a numerical summary of a collection of MassSpectrum objects.

summarySpectra(x, digits = 4)

Arguments

x
A list of MassSpectrum objects.

digits
Integer indicating the number of decimal places to be used.

Details

For each MassSpectrum on the list this function provides summary statistics of m/z points and signal intensities (number, minimum, mean, standard deviation, median, mean absolute deviation, maximum).

Value

A data.frame containing summary information of a collection of MassSpectrum objects.

Examples

# Load example data

data(spectra) # list of MassSpectra class objects
data(type)    # metadata

# Summary of spectra features (results for 20 first mass spectra)

summarySpectra(spectra[1:20])

#>           ID No.MZ  Min.MZ   Max.MZ Min.Int  Mean.Int   Std.Int Med.Int
#> 1  160408F21  1857 2500.05 12995.77      19  722.2649  572.9375     664
#> 2  160408F22  1857 2500.05 12995.77      18  637.1357  531.1853     582
#> 3  160408F23  1857 2500.05 12995.77       8  542.4976  447.7237     483
#> 4  160408F24  1857 2500.05 12995.77      34  957.3430  737.8402     903
#> 5  160408G01  1857 2500.05 12995.77      42 1129.8185  871.5639    1064
#> 6  160408G02  1857 2500.05 12995.77      24  637.1648  530.9029     595
#> 7  160408G03  1857 2500.05 12995.77      12  505.9499  442.6567     466
#> 8  160408G04  1857 2500.05 12995.77      43  984.1179  754.5525     936
#> 9  160408G05  1857 2500.05 12995.77      27  906.9138  686.9536     854
#> 10 160408G06  1857 2500.05 12995.77      46  906.9133  689.0228     891
#> 11 160408G07  1857 2500.05 12995.77       9  383.5385  339.6962     352
#> 12 160408G08  1857 2500.05 12995.77      26  756.7415  582.2340     715
#> 13 160408G09  1857 2500.05 12995.77      14  580.6947  522.7782     524
#> 14 160408G10  1857 2500.05 12995.77      30  730.3985  578.6044     686
#> 15 160408G11  1857 2500.05 12995.77      26  771.9882  651.0067     712
#> 16 160408G12  1857 2500.05 12995.77      86  995.9386  639.7584    1000
#> 17 160408G13  1857 2500.05 12995.77     128 1894.1125 1228.2008    1887
#> 18 160408G14  1857 2500.05 12995.77      27  925.0722  739.1213     866
#> 19 160408G15  1857 2500.05 12995.77      46  728.0070  615.4929     680
#> 20 160408G16  1857 2500.05 12995.77      37  932.3393  703.2897     897
#>     MAD.Int Max.Int
#> 1  486.2928    5888
#> 2  382.5108    6128
#> 3  401.7846    4713
#> 4  587.1096    7843
#> 5  652.3440    7860
#> 6  358.7892    5459
#> 7  298.0026    5198
#> 8  530.7708    7333
#> 9  558.9402    6068
#> 10 481.8450    7505
#> 11 237.2160    3444
#> 12 426.9888    5410
#> 13 343.9632    5480
#> 14 415.1280    5948
#> 15 441.8148    6724
#> 16 471.4668    6310
#> 17 944.4162   10911
#> 18 541.1490    7509
#> 19 373.6152    8147
#> 20 493.7058    7631

# Some pre-processing

sc.results <- screenSpectra(spectra, meta = type)

spectra <- sc.results$fspectra
type <- sc.results$fmeta

spectra <- transformIntensity(spectra, method = "sqrt")
spectra <- wavSmoothing(spectra)
spectra <- removeBaseline(spectra)

names(spectra) <- type$SpectID # spectra IDs are lost with removeBaseline()

# Summary of spectra features (results for positions 10 to 20)

summarySpectra(spectra[10:20])

#>           ID No.MZ  Min.MZ   Max.MZ Min.Int Mean.Int Std.Int Med.Int MAD.Int
#> 1  160408G06  1857 2500.05 12995.77       0   9.2793  7.2744  8.5795  4.6772
#> 2  160408G07  1857 2500.05 12995.77       0   5.8767  5.4660  5.2470  3.8580
#> 3  160408G08  1857 2500.05 12995.77       0   8.0786  7.0385  6.7555  5.0173
#> 4  160408G09  1857 2500.05 12995.77       0   7.5583  7.0469  6.5943  4.7602
#> 5  160408G10  1857 2500.05 12995.77       0   8.3169  7.2127  7.0014  5.2002
#> 6  160408G11  1857 2500.05 12995.77       0   8.4901  7.8110  6.8286  5.3824
#> 7  160408G12  1857 2500.05 12995.77       0   8.5018  6.3926  8.0312  4.4920
#> 8  160408G13  1857 2500.05 12995.77       0  11.3623  9.1872 10.0677  6.5494
#> 9  160408G14  1857 2500.05 12995.77       0   9.1624  8.1827  8.0785  5.9631
#> 10 160408G15  1857 2500.05 12995.77       0   8.4055  7.2018  7.4998  4.7401
#> 11 160408G16  1857 2500.05 12995.77       0   9.2486  7.6595  8.1129  5.5344
#>    Max.Int
#> 1  69.6900
#> 2  45.2214
#> 3  57.0512
#> 4  58.4061
#> 5  61.7756
#> 6  67.0592
#> 7  59.4995
#> 8  79.7723
#> 9  71.2000
#> 10 74.2152
#> 11 69.6843