Numerical summary of screenSpectra objects

This is a summary method for scSpectra objects that generates a numerical summary of the settings and results from applying screenSpectra to identify potential faulty, low-quality raw mass spectra.

# S3 method for scSpectra
summary(object, ncases = 10, ...)

Arguments

object
A scSpectra object as generated from screenSpectra.

ncases
Number of cases shown in the results table.

...
Other arguments.

Details

A table is generated that includes details of the numerical estimations along with mass spectra ID, A score and the label for each mass spectra, either potentially low-quality (failure) or good-quality (success).

Value

No return value, text printed on console.

See also

Examples

# Load example data

data(spectra) # list of MassSpectra objects

sc.results <- screenSpectra(spectra)
summary(sc.results)

#> (10 first mass spectra) 
#>           ID   A score   Class
#> 1  160408F21 0.1683850 success
#> 2  160408F22 0.1716699 success
#> 3  160408F23 0.1887859 success
#> 4  160408F24 0.1478541 success
#> 5  160408G01 0.1479799 success
#> 6  160408G02 0.1734508 success
#> 7  160408G03 0.1772491 success
#> 8  160408G04 0.1534341 success
#> 9  160408G05 0.1716697 success
#> 10 160408G06 0.1538116 success
#> 
#> ----------------------------
#> 
#> Scale estimator: Q 
#> Method: adj.boxplot 
#> Threshold: 1.5 
#> Limits: [0.1196,0.2942] 
#> Deriv. order: 1 
#> Lambda: 0.5 
#> No. potentially faulty spectra: 19 (4.75 %)