Summary of mass peak profiles

This function generates a numerical summary of a collection of MassPeaks objects.

summaryPeaks(x, digits = 4)

Arguments

x
A list of MassPeaks objects.

digits
Integer indicating the number of decimal places to be used.

Details

For each MassPeaks on the list this function provides summary statistics of m/z points, peak intensities and SNR thresholds (number, minimum, mean, standard deviation, median, mean absolute deviation, maximum).

Value

A data.frame containing summary information of a collection of MassPeaks objects.

Examples

# Load example data

data(spectra) # list of MassSpectra class objects
data(type)    # metadata

# Some pre-processing

sc.results <- screenSpectra(spectra, meta = type)

spectra <- sc.results$fspectra
type <- sc.results$fmeta

spectra <- transformIntensity(spectra, method = "sqrt")
spectra <- wavSmoothing(spectra)
spectra <- removeBaseline(spectra)
peaks <- detectPeaks(spectra)

names(peaks) <- type$SpectID # spectra IDs are lost after removeBaseline()

# Summary of peak profile features (results for positions 10 to 20)

summaryPeaks(peaks[10:20])

#>           ID No.MZ  Min.MZ   Max.MZ Min.Int Mean.Int Std.Int Med.Int MAD.Int
#> 1  160408G06    19 2834.57 12175.49 10.1084  28.6467 18.2885 21.1420 14.2866
#> 2  160408G07    15 3174.54 10501.83  7.7203  22.7226 12.9644 17.6168 14.3619
#> 3  160408G08    15 3170.65 10501.83 11.5025  28.6847 14.1637 27.9683 19.2894
#> 4  160408G09    14 3174.54 10501.83  9.5581  30.6943 16.5765 28.7717 20.3452
#> 5  160408G10    14 3174.54 10501.83 13.8665  30.4962 14.3382 30.1252 18.1878
#> 6  160408G11    15 2834.57 10501.83 13.7654  33.2247 17.1682 31.2916 22.2767
#> 7  160408G12    17 3166.77 12175.49 10.5198  27.4917 16.4809 18.1243 10.1326
#> 8  160408G13    17 2834.57 12175.49 13.5931  36.0895 20.1844 31.5853 22.2787
#> 9  160408G14    15 2834.57 10501.83 12.9963  33.5270 17.4889 33.7262 25.4982
#> 10 160408G15    17 3182.31 12175.49 11.1976  31.6522 18.8827 24.5131 13.6015
#> 11 160408G16    16 2834.57 10501.83 11.4986  31.3445 17.0594 27.9781 20.3878
#>    Max.Int Min.SNR Mean.SNR Std.SNR Med.SNR MAD.SNR Max.SNR
#> 1  69.6900  2.1612   6.1248  3.9102  4.5202  3.0545 14.9000
#> 2  45.2214  2.0011   5.8898  3.3604  4.5663  3.7226 11.7215
#> 3  57.0512  2.2926   5.7171  2.8230  5.5744  3.8446 11.3709
#> 4  58.4061  2.0079   6.4481  3.4823  6.0442  4.2740 12.2696
#> 5  61.7756  2.6666   5.8645  2.7573  5.7931  3.4975 11.8796
#> 6  67.0592  2.5575   6.1729  3.1897  5.8137  4.1388 12.4591
#> 7  59.4995  2.3419   6.1201  3.6689  4.0348  2.2557 13.2456
#> 8  79.7723  2.0755   5.5104  3.0819  4.8226  3.4016 12.1801
#> 9  71.2000  2.1795   5.6224  2.9328  5.6558  4.2760 11.9401
#> 10 74.2152  2.3623   6.6776  3.9836  5.1715  2.8695 15.6570
#> 11 69.6843  2.0777   5.6636  3.0824  5.0553  3.6838 12.5911